% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCI.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{bootstrapCI}
\usage{
bootstrapCI(data, n, alpha, border = NULL, col = NA)
}
\arguments{
\item{data}{the data from which the estimate is to be computed. For the default method a numeric vector}

\item{n}{integer:the number of replications}

\item{alpha}{significance level}

\item{border}{the color to draw the border of the confidence band,Use border = NA to omit borders}

\item{col}{the color for filling the confidence band,The default,NA}
}
\value{
a matrix for storing the confidence interval for the density estimation \code{n}
}
\description{
compute the confidence interval of kernel density estimation through bootstrap method.
}
\examples{
\dontrun{
bootstrapCI(faithful$eruptions,10000,0.05,border = TRUE,col="blue")
}
}
