% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-nearest.R
\name{knnde}
\alias{knnde}
\title{knnde}
\usage{
knnde(data, k, dim, col = NA)
}
\arguments{
\item{data}{the data from which the estimate is to be computed. For the default method a numeric vector}

\item{k}{the maximum number of nearest neighbors to search. The default value is set to 10.}

\item{dim}{the dimension of data}

\item{col}{the color(s) of the surface facets.}
}
\value{
the estimated kernel density \code{n}
}
\description{
the k-nearest neighbor density estimation
}
\examples{
\dontrun{
fit_knnde2 <- knnde(faithful, k=5, dim =2,col='red')
fit_knnde1 <- knnde(faithful$eruptions, k=5, dim =1)
}
}
