% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ryan-Joiner-test.R
\name{RJtest}
\alias{RJtest}
\title{Ryan-Joiner-test}
\usage{
RJtest(x, alpha)
}
\arguments{
\item{x}{the data from which the estimate is to be computed. For the default method a numeric vector}

\item{alpha}{significance level}
}
\value{
Ryan-Joiner test statistic,marginal value,the result of normality test\code{n}
}
\description{
Ryan-Joiner test for normality test
}
\examples{
\dontrun{
x <- rpois(100,2)
RJtest(x,0.01)
y <- rnorm(1000)
RJtest(y,0.01)
}
}
