% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwc}
\alias{rwc}
\title{Implement a random walk Metropolis sampler.}
\usage{
rwc(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of a normal distribution}

\item{x0}{the initial value of a random walk sequence}

\item{N}{the length of a random walk sequence}
}
\value{
a random walk Metropolis sequence and number of candidate points rejected \code{n}
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
set.seed(1234)
N = 2000
sigma = 1
x0 = 25
rw = rwc(sigma,x0,N)
}
}
